+++ test/Feature/CleanRegex/Replace/first/ReplacePatternTest.php
@@ -103,16 +103,4 @@ class ReplacePatternTest extends TestCase
                 return '';
             });
     }
-
-    /**
-     * @test
-     */
-    public function shouldReturn_nonReplacedStrategy()
-    {
-        // when
-        $result = pattern('Foo')->replace('Bar')->first()->otherwiseReturning('otherwise')->with('XXX');
-
-        // then
-        $this->assertSame('otherwise', $result);
-    }
 }